requirements.txt
Flask==2.3.2
gunicorn==20.1.0
flask-cors==4.0.1
requests==2.31.0
python-telegram-bot==21.5